requirements.txt
Pillow==9.2.0
Django==4.1
fpdf2==2.5.7